Moves locator classes to rsocket package
