+++ rsocket-routing-broker-spring/src/main/java/io/rsocket/routing/broker/spring/BrokerAutoConfiguration.java
@@ -31,11 +31,11 @@ import io.rsocket.routing.broker.RSocketIndex;
 import io.rsocket.routing.broker.RoutingTable;
 import io.rsocket.routing.broker.acceptor.BrokerSocketAcceptor;
 import io.rsocket.routing.broker.acceptor.ClusterSocketAcceptor;
-import io.rsocket.routing.broker.locator.CompositeRSocketLocator;
-import io.rsocket.routing.broker.locator.RSocketLocator;
-import io.rsocket.routing.broker.locator.UnicastRSocketLocator;
-import io.rsocket.routing.broker.locator.UnicastRSocketLocator.Loadbalancer;
-import io.rsocket.routing.broker.locator.WeightedStatsAwareRSocket;
+import io.rsocket.routing.broker.rsocket.CompositeRSocketLocator;
+import io.rsocket.routing.broker.rsocket.RSocketLocator;
+import io.rsocket.routing.broker.rsocket.UnicastRSocketLocator;
+import io.rsocket.routing.broker.rsocket.UnicastRSocketLocator.Loadbalancer;
+import io.rsocket.routing.broker.rsocket.WeightedStatsAwareRSocket;
 import io.rsocket.routing.broker.query.CombinedRSocketQuery;
 import io.rsocket.routing.broker.query.RSocketQuery;
 import io.rsocket.routing.broker.rsocket.MulticastRSocketLocator;
